Link Windows folders to the Linux home directory in WSL
Changes:
- Added conditional check for the existence of the user-specific directory in WSL
- Linked the `$USERPROFILE/Documents` directory to the home directory if the user-specific directory doesn't exist
- Changed the symlink creation command for the `winHome` symlink to exclude the trailing slash in the source path

+++ code/link_files.py
@@ -99,11 +99,19 @@ class FileLinker(Script):
         if self._is_wsl() and os.path.exists("/mnt/c/Users"):
             link_cmds = [
                 f"ln -nfs $USERPROFILE/Downloads $HOME/",
-                f"ln -nfs $USERPROFILE/Dropbox/Documents $HOME/",
                 f"ln -nfs $USERPROFILE/Videos $HOME/",
             ]
+            if os.path.exists("/mnt/c/Users/mjo97"):
+                link_cmds += [
+                    f"ln -nfs $USERPROFILE/Dropbox/Documents $HOME/",
+                ]
+            else:
+                link_cmds += [
+                    f"ln -nfs $USERPROFILE/Documents $HOME/",
+                ]
+
             if not os.path.islink(f"{HOME}/winHome"):
-                link_cmds.append(f"ln -fs -T $USERPROFILE/ {HOME}/winHome")
+                link_cmds.append(f"ln -fs -T $USERPROFILE {HOME}/winHome")
 
             link_cmds = [f"source {WIN_CONV_PATH}; " + it for it in link_cmds]
             self.shell.exec_list("Symbolic linking windows folders", *link_cmds)
